Move the post-load data extraction/massage into repo
